:truck:rename: api 폴더 lib 폴더로 이동
